ui: fix + improve fields
This change adds the missing field for transaction count. It also populates the missing connection count value in the gRPC-web request (even though transaction count was the value missing a field?? Dynamically-typed languages truly horrify me). 

I also tweaked some (but not all) of the default values to be in line with the tm-load-test CLI's defaults.

Updates https://github.com/orijtech/cosmosloadtester/issues/8